+++ packages/@expo/prebuild-config/src/plugins/unversioned/expo-splash-screen/withIosSplashAssets.ts
@@ -8,7 +8,7 @@ import path from 'path';
 import { IOSSplashConfig } from './getIosSplashConfig';
 import {
   ContentsJsonImage,
-  ContentsJsonImageAppearance,
+  ContentsJsonAppearance,
   createContentsJsonItem,
   writeContentsJsonAsync,
 } from '../../icons/AssetContents';
@@ -111,43 +111,28 @@ async function copyImageFiles({
   darkTabletImage?: string | null;
   logoWidth: number;
 }) {
-  await Promise.all(
-    [
-      { ratio: 1, suffix: '' },
-      { ratio: 2, suffix: '@2x' },
-      { ratio: 3, suffix: '@3x' },
-    ].map(async ({ ratio, suffix }) => {
-      const filePath = path.resolve(
-        iosNamedProjectRoot,
-        IMAGESET_PATH,
-        `${PNG_FILENAME}${suffix}.png`
-      );
-
-      const size = logoWidth * ratio;
-
-      const { source } = await generateImageAsync(
-        { projectRoot, cacheType: IMAGE_CACHE_NAME },
-        {
-          src: image,
-          width: size,
-          height: size,
-          resizeMode: 'contain',
-        }
-      );
-      return await fs.writeFile(filePath, source);
-    })
-  );
-
   await generateImagesAssetsAsync({
     async generateImageAsset(item, fileName) {
-      // Using this method will cache the images in `.expo` based on the properties used to generate them.
-      // this method also supports remote URLs and using the global sharp instance.
-      const { source } = await generateImageAsync({ projectRoot, cacheType: IMAGE_CACHE_NAME }, {
-        src: item,
-      } as any);
-      // Write image buffer to the file system.
-      // const assetPath = join(iosNamedProjectRoot, IMAGESET_PATH, filename);
-      await fs.writeFile(path.resolve(iosNamedProjectRoot, IMAGESET_PATH, fileName), source);
+      [
+        { ratio: 1, suffix: '' },
+        { ratio: 2, suffix: '@2x' },
+        { ratio: 3, suffix: '@3x' },
+      ].map(async ({ ratio, suffix }) => {
+        const size = logoWidth * ratio;
+        // Using this method will cache the images in `.expo` based on the properties used to generate them.
+        // this method also supports remote URLs and using the global sharp instance.
+        const { source } = await generateImageAsync({ projectRoot, cacheType: IMAGE_CACHE_NAME }, {
+          src: item,
+          width: size,
+          height: size,
+        } as any);
+        // Write image buffer to the file system.
+        // const assetPath = join(iosNamedProjectRoot, IMAGESET_PATH, filename);
+        await fs.writeFile(
+          path.resolve(iosNamedProjectRoot, IMAGESET_PATH, `${fileName}${suffix}.png`),
+          source
+        );
+      });
     },
     anyItem: image,
     darkItem: darkImage,
@@ -179,11 +164,11 @@ async function generateImagesAssetsAsync({
   await Promise.all(items.map(([item, fileName]) => generateImageAsset(item, fileName)));
 }
 
-const darkAppearances: ContentsJsonImageAppearance[] = [
+const darkAppearances: ContentsJsonAppearance[] = [
   {
     appearance: 'luminosity',
     value: 'dark',
-  } as ContentsJsonImageAppearance,
+  } as ContentsJsonAppearance,
 ];
 
 export function buildContentsJsonImages({
@@ -220,43 +205,48 @@ export function buildContentsJsonImages({
         idiom: 'universal',
         appearances: darkAppearances,
         scale: '1x',
+        filename: `${darkImage}.png`,
       }),
     darkImage &&
       createContentsJsonItem({
         idiom: 'universal',
         appearances: darkAppearances,
         scale: '2x',
+        filename: `${darkImage}@2x.png`,
       }),
     darkImage &&
       createContentsJsonItem({
         idiom: 'universal',
         appearances: darkAppearances,
         scale: '3x',
+        filename: `${darkImage}@3x.png`,
       }),
     // Tablet light
     tabletImage &&
       createContentsJsonItem({
         idiom: 'ipad',
-        filename: tabletImage,
+        filename: `${tabletImage}.png`,
         scale: '1x',
       }),
     tabletImage &&
       createContentsJsonItem({
         idiom: 'ipad',
         scale: '2x',
+        filename: `${tabletImage}@2x.png`,
       }),
     // Phone dark
     darkTabletImage &&
       createContentsJsonItem({
         idiom: 'ipad',
         appearances: darkAppearances,
-        filename: darkTabletImage ?? undefined,
+        filename: `${darkTabletImage}.png`,
         scale: '1x',
       }),
     darkTabletImage &&
       createContentsJsonItem({
         idiom: 'ipad',
         appearances: darkAppearances,
+        filename: `${darkTabletImage}@2x.png`,
         scale: '2x',
       }),
   ].filter(Boolean) as ContentsJsonImage[];