+++ packages/expo-router/src/ui/TabTrigger.tsx
@@ -18,13 +18,22 @@ type PressablePropsWithoutFunctionChildren = Omit<PressableProps, 'children'> & 
 };
 
 export type TabTriggerProps = PressablePropsWithoutFunctionChildren & {
-  /** Name of tab. When used within a `<TabList />` this sets the name of the tab. Otherwise, this references the name. */
+  /**
+   *  Name of tab. When used within a `TabList` this sets the name of the tab.
+   * Otherwise, this references the name.
+   */
   name: string;
-  /** Name of tab. Required when used within a `<TabList />` */
+  /**
+   * Name of tab. Required when used within a `TabList`.
+   */
   href?: Href;
-  /** Forward props to child component. Useful for custom wrappers. */
+  /**
+   * Forward props to child component. Useful for custom wrappers.
+   */
   asChild?: boolean;
-  /** Reset the route when switching to the tab */
+  /**
+   * Resets the route when switching to a tab.
+   */
   reset?: SwitchToOptions['reset'] | 'onLongPress';
 };
 
@@ -42,12 +51,14 @@ export type TabTriggerSlotProps = PressablePropsWithoutFunctionChildren &
 const TabTriggerSlot = Slot as React.ForwardRefExoticComponent<TabTriggerSlotProps>;
 
 /**
- * Creates a trigger to navigate to a tab. `<TabTrigger />` functionality slightly changes when used as a child of `<TabList />`. In this instance, the `href` prop is required, and the trigger also defines what routes are present in the `<Tabs />`.
+ * Creates a trigger to navigate to a tab. When used as child of `TabList`, its
+ * functionality slightly changes since the `href` prop is required,
+ * and the trigger also defines what routes are present in the `Tabs`.
  *
- * When used outside of `<TabList />`, `<TabTrigger />` no longer requires a `href`.
+ * When used outside of `TabList`, this component no longer requires an `href`.
  *
  * @example
- * ```ts
+ * ```tsx
  * <Tabs>
  *  <TabSlot />
  *  <TabList>
@@ -99,7 +110,9 @@ export function isTabTrigger(
  * Options for `switchTab` function.
  */
 export type SwitchToOptions = {
-  /** Navigate and reset the history */
+  /**
+   * Navigate and reset the history.
+   */
   reset?: ExpoTabsResetValue;
 };
 
@@ -123,7 +136,7 @@ export type TriggerProps = {
 };
 
 /**
- * Utility hook creating custom `<TabTrigger />`
+ * Utility hook creating custom `TabTrigger`.
  */
 export function useTabTrigger(options: TabTriggerProps): UseTabTriggerResult {
   const { state, navigation } = useNavigatorContext();