+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoModule.kt
@@ -16,6 +16,7 @@ import expo.modules.kotlin.functions.Coroutine
 import expo.modules.kotlin.modules.Module
 import expo.modules.kotlin.modules.ModuleDefinition
 import expo.modules.kotlin.types.Either
+import expo.modules.video.enums.AudioMixingMode
 import expo.modules.video.enums.ContentFit
 import expo.modules.video.player.VideoPlayer
 import expo.modules.video.records.BufferOptions
@@ -316,6 +317,16 @@ class VideoModule : Module() {
         }
         .set { ref: VideoPlayer, intervalSeconds: Float ->
           ref.intervalUpdateClock.interval = (intervalSeconds * 1000).toLong()
+        }
+
+      Property("audioMixingMode")
+        .get { ref: VideoPlayer ->
+          ref.audioMixingMode
+        }
+        .set { ref: VideoPlayer, audioMixingMode: AudioMixingMode ->
+          appContext.mainQueue.launch {
+            ref.audioMixingMode = audioMixingMode
+          }
         }
 
       Function("replace") { ref: VideoPlayer, source: Either<Uri, VideoSource>? ->