+++ packages/expo-video/android/src/main/java/expo/modules/video/player/VideoPlayerListener.kt
@@ -3,6 +3,7 @@ package expo.modules.video.player
 import androidx.annotation.OptIn
 import androidx.media3.common.Tracks
 import androidx.media3.common.util.UnstableApi
+import expo.modules.video.enums.AudioMixingMode
 import expo.modules.video.enums.PlayerStatus
 import expo.modules.video.records.PlaybackError
 import expo.modules.video.records.VideoSource
@@ -19,4 +20,5 @@ interface VideoPlayerListener {
   fun onTracksChanged(player: VideoPlayer, tracks: Tracks) {}
   fun onTimeUpdate(player: VideoPlayer, timeUpdate: TimeUpdate) {}
   fun onPlayedToEnd(player: VideoPlayer) {}
+  fun onAudioMixingModeChanged(player: VideoPlayer, audioMixingMode: AudioMixingMode, oldAudioMixingMode: AudioMixingMode?) {}
 }