+++ packages/@expo/server/CHANGELOG.md
@@ -10,6 +10,12 @@
 
 ### 💡 Others
 
+## 0.5.0-preview.1 — 2024-11-05
+
+### 🎉 New features
+
+- Add `expo-router/rsc/headers` for accessing request headers in server components. ([#32099](https://github.com/expo/expo/pull/32099) by [@EvanBacon](https://github.com/EvanBacon))
+
 ## 0.5.0-preview.0 — 2024-10-22
 
 ### 🎉 New features